+++ app/setup/component/etcdServerInstallerStep_internal_test.go
@@ -14,7 +14,7 @@ const etcdServerResourceURL = "https://url.server.com/etcd/resource.yaml"
 
 var etcdServerSetupCtx = &ctx.SetupContext{
 	AppVersion: "1.2.3",
-	AppConfig: ctx.Config{
+	AppConfig: &ctx.Config{
 		TargetNamespace:       testTargetNamespaceName,
 		EtcdServerResourceURL: etcdServerResourceURL,
 	},